feat: validate connection
Add a task to verify that the command has triggered a data point update.
If no data update has been done the device is in a limbo state and does not
do anything. It resets successfully if an update call is executed then
which is happening in the task after a second of waiting if no new data
has been received.